Check for Balanced Tree
Given a binary tree, find if it is height balanced or not. 
A tree is height balanced if difference between heights of left and right subtrees is not more than one for all nodes of tree. 
Constraints:
1 <= Number of nodes <= 10^5
0 <= Data of a node <= 10^6
Expected time complexity: O(N)
Expected auxiliary space: O(h) , where h = height of tree