Binary Search in forest
There are n trees in a forest. The heights of the trees is stored in array tree[], where tree[i] denotes the height of the ith tree in the forest. If the ith tree is cut at a height H, then the wood collected is tree[i] - H, provided tree[i] > H. If the total woods that needs to be collected is exactly equal to k, find the height H at which every tree should be cut (all trees have to be cut at the same height). If it is not possible then return -1 else return H.
Expected Time Complexity: O(n log h)
Expected Auxiliary Space: O(1)
Constraints: 
1 <= n <= 10^4
1 <= tree[i] <= 10^3
1 <= k <= 10^7