Sort an array according to the other
Given two integer arrays A1[ ] and A2[ ] of size N and M respectively. Sort the first array A1[ ] such that all the relative positions of the elements in the first array are the same as the elements in the second array A2[ ].
See example for better understanding.
Note: If elements are repeated in the second array, consider their first occurrence only.
For the elements not present in A2[ ] but in A1[ ], it appends them at the last in increasing order. 
Expected Time Complexity: O(N * Log(N)).
Expected Auxiliary Space: O(N).
Constraints:
1 ≤ N, M ≤ 10^6
1 ≤ A1[i], A2[i] ≤ 10^6